requirements.txt
ftfy 
regex 
git+https://github.com/openai/CLIP.git
hydra-core==1.3.2
matplotlib==3.10.1
numpy==2.2.5
omegaconf==2.3.0
pandas==2.2.3
Pillow==11.2.1
psutil
scikit_learn==1.6.1
seaborn==0.13.2
timm==1.0.15
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers
codecarbon